requirements.txt
requests>=2.32.0
google-generativeai>=0.8.0
